chore: extract generateTokenId to shared module and add CLAUDE.md
Move generateTokenId() from combat/combatUtils.ts to shared/idUtils.ts
to decouple it from combat-specific code, enabling parallel development
across feature branches. Add project-level CLAUDE.md with architecture
notes, branch protection policy, and documentation conventions.
