+++ src/dock/BottomDock.tsx
@@ -3,7 +3,7 @@ import * as Y from 'yjs'
 import type { Scene } from '../yjs/useScenes'
 import type { MapToken, Entity, Blueprint } from '../shared/entityTypes'
 import { defaultNPCPermissions } from '../shared/permissions'
-import { generateTokenId } from '../combat/combatUtils'
+import { generateTokenId } from '../shared/idUtils'
 import { nextNpcName } from '../shared/characterUtils'
 import { MapDockTab } from './MapDockTab'
 import { TokenDockTab } from './TokenDockTab'
