+++ src/shared/idUtils.ts
@@ -1,0 +1,5 @@
+export function generateTokenId(): string {
+  return (
+    self.crypto?.randomUUID?.() ?? Math.random().toString(36).slice(2) + Date.now().toString(36)
+  )
+}
